Arduino-ros for wheel odometry
This code,written by Shaheer, allows for the pose to be published by getting information from the Arduino. 
In case this breaks, this was originally shared with me along with the workspace, unit testing was complete (i.e. the code is functional when the arduino is connected to wheel encoder circuit), integrated testing is required (i.e. when the entire subsystems and larger code base are connected).